Changed type to float

+++ ts4uc/agents/a3c/train.py
@@ -191,7 +191,7 @@ if __name__ == "__main__":
     parser.add_argument('--num_nodes', type=int, required=False, default=32)
     parser.add_argument('--entropy_coef', type=float, required=False, default=0.01)
     parser.add_argument('--forecast_horizon_hrs', type=int, required=False, default=12)
-    parser.add_argument('--credit_assignment_1hr', type=int, required=False, default=0.9)
+    parser.add_argument('--credit_assignment_1hr', type=float, required=False, default=0.9)
     parser.add_argument('--minibatch_size', type=int, required=False, default=None)
     parser.add_argument('--update_epochs', type=int, required=False, default=4)
     parser.add_argument('--observation_processor', type=str, required=False, default='LimitedHorizonProcessor')
